Added suspension and banning functionality
Added functionality to "admin_page.php" such that users that don't have admin permissions can be suspended/banned from the website. If a suspended/banned user tries to login, they will get an error message.
Also added logout buttons to the admin and user pages that redirect to "index.php", as well as adjusted "layout.css" such that the background of webpages doesn't move with scrolling.